//Cambios v2.2 1. Se crea tabla CxcVentas y CxcPagos para el manejo de los métodos de pago a nivel de base de datos y código 2. Se organiza los diferentes opciones que se tienen para métodos de pago y detalles de venta. 3. Se organiza en la vista de ordenes de servicio para que en histórico se visualize los botones dependiendo del estado de la orden.

+++ Login/Controllers/PedidoController.cs
@@ -149,7 +149,7 @@ namespace Plataforma.Controllers
                 return RedirectToAction("AgregarFactura");
             }
             //decimal iva = subtotal * 0.19M;
-            decimal total = venta.Pedidos.Sum(p => p.SubTotal);
+            decimal total = (decimal)venta.Pedidos.Sum(p => p.SubTotal);
             decimal totalIva = ((total * Iva) / 100);
             if (totalIva > 0)
             {
